+++ lab2/Parsing.py
@@ -3,75 +3,7 @@ import requests
 from bs4 import BeautifulSoup
 import csv
 
-
-INSTRUMENTS: list[str] = ['trumpet', 'ukulele', 'harp']
 BASE_URL: str = "https://mixkit.co/free-stock-music/instrument/"
-
-
-class FilePathIterator:
-    """Итератор по путям файлов из CSV-аннотации.
-
-    Позволяет перебирать файлы, сохранённые в CSV-файл с колонками:
-    Filename, Absolute Path, Relative Path.
-
-    Attributes:
-        annotation_file: Путь к CSV-файлу аннотации.
-        rows: Список строк из CSV-файла.
-        _index: Текущий индекс при переборе.
-
-    """
-
-    def __init__(self, annotation_file: str) -> None:
-        """
-        Инициализация итератора.
-
-        :param annotation_file: Путь к CSV-файлу аннотации.
-        :raises FileNotFoundError: Если файл не существует.
-        """
-        self.annotation_file = Path(annotation_file)
-        self.rows: list[dict] = []
-
-        if not self.annotation_file.exists():
-            raise FileNotFoundError(
-                f"Аннотация не найдена: {annotation_file}"
-            )
-
-        try:
-            with open(
-                self.annotation_file, 'r', encoding='utf-8'
-            ) as f:
-                reader = csv.DictReader(f)
-                self.rows = list(reader)
-        except Exception as e:
-            print(f"Ошибка чтения аннотации {annotation_file}: {e}")
-            self.rows = []
-
-        self._index: int = 0
-
-    def __iter__(self):
-        """Начало итерации."""
-        self._index = 0
-        return self
-
-    def __next__(self) -> tuple[str, str, str]:
-        """
-        Получить следующую строку.
-
-        :return: Кортеж (filename, abs_path, rel_path).
-        :raises StopIteration: Когда не осталось строк.
-        """
-        if self._index >= len(self.rows):
-            raise StopIteration
-
-        row = self.rows[self._index]
-        self._index += 1
-
-        return (
-            row.get('Filename', ''),
-            row.get('Absolute Path', ''),
-            row.get('Relative Path', '')
-        )
-
 
 def fetch_audio_urls(instrument: str) -> list[str]:
     """
@@ -138,7 +70,6 @@ def fetch_audio_urls(instrument: str) -> list[str]:
         print(f"Ошибка парсинга страницы для {instrument}: {e}")
         return []
 
-
 def download_audio_files(
     instrument: str,
     urls: list[str],
@@ -190,7 +121,6 @@ def download_audio_files(
     print(f"Все аудио обработаны для инструмента '{instrument}'")
     return data
 
-
 def save_annotation(
     data: list[list[str]],
     csv_path: Path
@@ -212,78 +142,3 @@ def save_annotation(
         print(f"Ошибка при сохранении CSV: {e}")
         raise
 
-def count(lst: list[list[str]]) -> int:
-    """Вычислить количество скаченных файлов.
-    :param lst: Список всех ссылок на скачивание.
-    """
-    count = 0
-    for sublist in lst:
-        if count == 0:
-            count = len(sublist)
-        else:
-            count = min(count, len(sublist))
-    return count
-
-def print_annotation(csv_path: Path) -> None:
-    """
-    Вывести содержимое аннотации через итератор.
-
-    :param csv_path: Путь к CSV-файлу.
-    """
-    print("\n--- Чтение аннотации через FilePathIterator ---")
-    try:
-        iterator = FilePathIterator(str(csv_path))
-        for filename, abs_path, rel_path in iterator:
-            print(f"Имя файла: {filename} Относительный путь: {rel_path}; Абсолютный путь: {abs_path}")
-    except Exception as e:
-        print(f"Ошибка итератора: {e}")
-
-
-def main() -> None:
-    """
-    Основная функция для скачивания аудиофайлов.
-
-    Последовательно:
-    1. Получает URL для каждого инструмента
-    2. Определяет максимальное количество файлов
-    3. Скачивает аудиофайлы
-    4. Сохраняет аннотацию в CSV
-    5. Выводит загруженные файлы
-    """
-    try:
-        script_dir = Path(__file__).parent
-
-        print("Получение URL аудиофайлов...")
-        audio_map: dict[str, list[str]] = {}
-        all_urls: list[list[str]] = []
-
-        for instrument in INSTRUMENTS:
-            urls = fetch_audio_urls(instrument)
-            audio_map[instrument] = urls
-            all_urls.append(urls)
-
-        count_audio = count(all_urls)
-        
-        print("Скачивание аудиофайлов...")
-        all_data: list[list[str]] = []
-        for instrument in INSTRUMENTS:
-            data = download_audio_files(
-                instrument,
-                audio_map[instrument],
-                count_audio,
-                script_dir
-            )
-            all_data.extend(data)
-
-        csv_path = script_dir / 'data.csv'
-        save_annotation(all_data, csv_path)
-
-        print_annotation(csv_path)
-
-    except Exception as e:
-        print(f"Критическая ошибка в main(): {e}")
-        raise
-
-
-if __name__ == "__main__":
-    main()
